### 2022-04-14 #4318 @arfon
Output of `cloc` for actual software:

```
Language                     files          blank        comment           code
Jupyter Notebook                35              0          11385           7388
Python                          29           1277            805           4789
Markdown                         4            146              0            267
YAML                             4             23              3            184
Bourne Shell                     5             25             13             60
Dockerfile                       1             20             10             59
JSON                             2              0              0             59
INI                              1              8              0             50
TOML                             1              5              2             50
DOS Batch                        1              8              1             26
make                             1             11              0             26
HTML                             2              0              0             18
reStructuredText                 3             11             45             14
SUM:                            89           1534          12264          12990
```
