### 2022-04-15 20060e7
Update paper.bib with DOIs (#20)
