+++ paper/paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={447--457},
   year={2015},
-  organization={Wiley Online Library}
+  organization={Wiley Online Library},
+  doi = {10.1145/3359141},
 }
 
 @article{wang2019how,
@@ -44,7 +45,8 @@
   booktitle={2021 IEEE International Conference on Image Processing (ICIP)},
   pages={2633--2637},
   year={2021},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/icip42928.2021.9506683}
 }
 
 @inproceedings{dutta2019via,
@@ -52,7 +54,8 @@
   author={Dutta, Abhishek and Zisserman, Andrew},
   booktitle={Proceedings of the 27th ACM international conference on multimedia},
   pages={2276--2279},
-  year={2019}
+  year={2019},
+  doi={10.1145/3343031.3350535}
 }
 
 @article{bernal2019gtcreator,
@@ -63,7 +66,8 @@
   number={2},
   pages={191--201},
   year={2019},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11548-018-1864-x}
 }
 
 @book{gamma1995design,
