### 2022-04-21 #4318 @danielskatz
In addition, if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).  The suggestions you made above are editors, and we generally don't ask editors to also be reviewers.
