### 2022-04-26 #4318 @matthewfeickert
> I wonder if you would be interested in reviewing this submission?

Thanks for the ping @danielskatz. I would be open to reviewing this, but I have been slacking in my JOSS responsibilities for the last month and need to finish the Asimov review, so I probably wouldn't be able to devote my full attention to this for probably at least a week or so.
