### 2022-04-26 #4318 @danielskatz
@matthewfeickert - your thoughts on the scholarly contribution criteria would be welcome. Please ask questions of the author here.  On the pro side, and the reason this didn't undergo a scope review is the 5000 lines of Python. It does appear that this was written with a small number of commits, but they were over 18 months, which might just indicate a non-typical style of using git.
