### 2022-04-26 #4318 @csadorf
> 👋 @csadorf & @a-slide - Would one of you be willing to review this JOSS submission?

Yes, I am available.
