+++ .github/workflows/docs_pages.yaml
@@ -28,7 +28,7 @@ jobs:
         cp -r _build/html/* ./gh-pages/
     - name: Deploy documentation
       if: ${{ github.event_name == 'push' }}
-      uses: JamesIves/github-pages-deploy-action@4.1.4
+      uses: JamesIves/github-pages-deploy-action@v4.3.3
       with:
         branch: gh-pages
         folder: gh-pages
