### 2022-04-27 #4318 @Athene-ai
> @Athene-ai - As a JOSS policy, as I wrote above:
>
> > we prefer to start new reviewers off with just one review initially
>
> Once your initial review is complete, we can assign you more

Oh I see :-) So I can not wait to start and complete my initial review  in order to contribute to review more
