### 2022-04-27 #4318 @ibayer
@matthewfeickert

> if Ipyannotator meets the [scholarly effort criteria for JOSS](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

Thanks for bringing this up. In addition to @itepifanio comment above I want to mention that we have been following a Private Development, Public Releases workflow[0].
I guess this makes an effort estimate based on number of commits difficult. Please find below a screenshot from our private development branch. I hope the number of commits are more in line with your expectations.
![image](https://user-images.githubusercontent.com/1497140/165515010-e4ab4388-fa21-43ab-bc4c-f28f24bb9931.png)

Also now that the project matures we are planning to also develop more code in the open.

[0] https://www.braintreepayments.com/blog/our-git-workflow/
