### 2022-04-27 #4318 @matthewfeickert
> Palaimon's uses Ipyannotator internally and since we work with RnD projects our iterations are pretty quickly with ideas that are not always being added to the library core. To keep the iterations flexible we use a private GitLab repository and release all changes at once on Github, that's why we have just a few commits.

@itepifanio @ibayer Thanks for clarifying. I didn't understand from the Issue or from the docs that this is actually a project from a company that is being made incrementally open. While different then most projects submitted, there's nothing wrong with this approach. :+1:

> Regarding the documentation, we always try the tutorial-first approach

That's fine for your team, but it appears it isn't "tutorial-first" but rather "tutorial-only". As it stands, your project has no API reference documentation (the [Internal Components](https://palaimon.github.io/ipyannotator/docs/internal-components.html) page is essentially just a view of the source code) and the docs/tutorial webpage is unversioned, so in terms of the [JOSS review criteria for API documentation](https://joss.readthedocs.io/en/latest/review_criteria.html#api-documentation) I would have to say this is not passing.

To be clear, I think user guides are great, but I think API documentation is necessary.
