### 2022-04-27 #4318 @ibayer
@matthewfeickert

> To be clear, I think user guides are great, but I think API documentation is necessary.

I understand your concern. ipyannotator is build using https://nbdev.fast.ai/ which is somewhat popular (3.2k stars) and is also being used to build the popular https://github.com/fastai/fastai library (22k stars). As is, nbdev is a bit special *literate programming style* and doesn't produce the typical sphinx type API docs. To be clear, I'm not opposing your view. I'm just trying to provide background to avoid the impression that the missing API docs are viewed as being an indication of us being lazy or sloppy. ;)

We'll discuss internally what we can propose to accommodate the need for generated API docs.
