### 2022-04-27 #4318 @itepifanio
@matthewfeickert we could provide an "API" page reference like the [adaptr](https://inceptdk.github.io/adaptr/reference/index.html) (one of the [latest papers accepted](https://joss.theoj.org/papers/10.21105/joss.04284)).

> doesn't produce the typical sphinx type API docs

Our API doesn't change much, so a manual reference could be an option.
