### 2022-04-28 #4318 @matthewfeickert
> I'm just trying to provide background to avoid the impression that the missing API docs are viewed as being an indication of us being lazy or sloppy

Yup. I fully appreciate that and I in no way view your team as lazy as I'm quite familiar with both literate programming and nbdev. `fast.ai` is great and Jeremy Howard is very skilled at API design &mdash; I use `ghapi` regularly, which I'll note [does have API reference docs](https://ghapi.fast.ai/core.html).

> Our API doesn't change much, so a manual reference could be an option.

Please don't. That's a bad option for everyone &mdash; it creates busy work for you and it will only lead to wrong APIs in the future.
