### 2022-05-06 50bb579
Bugfix build docs workflow
