### 2022-05-07 d43fafb
Bump abatilo/actions-poetry from 2.0.0 to 2.1.4 (#18)
Bumps [abatilo/actions-poetry](https://github.com/abatilo/actions-poetry) from 2.0.0 to 2.1.4.
- [Release notes](https://github.com/abatilo/actions-poetry/releases)
- [Changelog](https://github.com/abatilo/actions-poetry/blob/master/.releaserc)
- [Commits](https://github.com/abatilo/actions-poetry/compare/v2.0.0...v2.1.4)

---
updated-dependencies:
- dependency-name: abatilo/actions-poetry
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>

Co-authored-by:
[…]
