### 2022-05-07 2e7ea5b
Deploying to gh-pages from @ palaimon/ipyannotator@c9a002bcc314df38227a1b586912057032453a1e 🚀
