### 2022-05-09 #4318 @itepifanio
@matthewfeickert can you check [Ipyannotator docs](https://palaimon.github.io/ipyannotator/index.html) again? The new release contains the changes that you suggested.
