### 2022-06-10 #4318 @ibayer
@danielskatz Thanks for staying on top of things. Please let us know if we can do anything else to pass the [PRE REVIEW] stage and get going. We are very much looking forward to getting more feedback that helps us to improve ipyannotator. :)
