### 2022-06-15 #4318 @matthewfeickert
Thanks for everyone's extreme patience here.

> can you check [Ipyannotator docs](https://palaimon.github.io/ipyannotator/index.html) again? The new release contains the changes that you suggested.

Thanks @itepifanio. As `Annotator` only exposes 3 public methods this is a pretty good start, so we can discuss any further bits in the review proper.

> Also, repeating, if you can do this review (based on your interest and the discussion so far), it would be very helpful.

I'll support moving `ipyannotator` into review status. I'm willing to do the review, and I hope it can be done quickly, but I also know that things are going to be pretty hectic between now and the last week of July. I can fully appreciate that both the editors and the authors might (very reasonably!) require a more reasonable iteration speed then I can provide (2 weeks should generally more than enough time in most situations). So while I'm ready to move forward I also understand if the authors would be more comfortable with another reviewer and won't be offended if they request that I be replaced if I fail to stick to a reasonable time in the review. Everyone's time is valuable here and I don't want the authors time abused.

Okay, with that out of the way, onward to the review if @csadorf is still game! :rocket:

@danielskatz Also thanks very much for pinging me here. I really do appreciate it, especially when I'm fighting to keep any notifications on the board. Hopefully you won't have to keep doing it, but thanks for not giving up!

> congratulations on your new job!!

Thanks! I truly appreciate it. :) (One of the reasons things have been much busier than normal)
