### 2022-06-15 #4480 @editorialbot
[REVIEW]: Ipyannotator: the infinitely hackable annotation framework
**Submitting author:** @itepifanio (<a href="http://orcid.org/0000-0003-0279-0942">Ítalo Epifânio</a>)
**Repository:** https://github.com/palaimon/ipyannotator/
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.8.5
