### 2022-07-05 #4480 @ibayer
@danielskatz We have just been informed that an ipyannotator introduction talk, that @itepifanio have submitted to Python Nordeste, has been accepted.

It would be great if we could in cooperate feedback from this review before this talk. Please let me know if we can do something to get the review started.
