### 2022-07-08 #4480 @ibayer
Thanks for the feedback!

> The significance of contributions by the first author to the source code base cannot be verified.

We have been following a Private Development, Public Releases workflow[0]. The downside of this approach is that
the number of commits https://github.com/openjournals/joss-reviews/issues/4318#issuecomment-1110934810 as well as individual contributions can't be deduced from the git history.

All authors have contributed code, but  @itepifanio has written by far the majority. Maybe the [CHANGELOG.md](https://github.com/palaimon/ipyannotator/blob/main/CHANGELOG.md) can be used to support my assurance.

We'll address all other points in detail later.

[0] https://www.braintreepayments.com/blog/our-git-workflow/
