+++ README.md
@@ -2,7 +2,9 @@
 
 ![CI-Badge](https://github.com/palaimon/ipyannotator/workflows/CI/badge.svg)
 
-Ipyannotator is a flexible annotation system. The library contains some pre-defined annotators that can be used out of the box, but it also can be extend and customized according to the users needs.
+Ipyannotator is a flexible annotation system. The library contains some pre-defined annotators that can be used out of the box, but it also can be extend and customized according to the users needs. Check our [tutorials](https://palaimon.github.io/ipyannotator/docs/tutorials.html) for a quickly understanding of it's usage and check our [API](https://palaimon.github.io/ipyannotator/nbs/21_api_doc.html) for quick reference.
+
+This library has been written in the [literate programming style](https://en.wikipedia.org/wiki/Literate_programming) popularized for jupyter notebooks by [nbdev](https://www.fast.ai/2019/12/02/nbdev/). Further than our documentation the jupyter notebooks located at `nbs/` can be explored to learn more about the inner working of Ipyannotator.
 
 We hope this repository helps you to explore how annotation UI's can be quickly built using only python code and leveraging many awesome libraries ([ipywidgets](https://github.com/jupyter-widgets/ipywidgets), [voila](https://github.com/voila-dashboards/voila), [ipycanvas](https://github.com/martinRenou/ipycanvas), etc.) from the [jupyter Eco-system](https://jupyter.org/).
 
@@ -14,58 +16,49 @@ Please let us know if you find this repository useful. Your feedback will help u
 
 ## Install
 
+Ipyannotator is available on Pypi and can be installed using:
+
 `pip install ipyannotator`
 
-**dependencies (should be handled by pip)**
+## Running ipyannotator
 
-```
-python = "^3.7"
-traitlets = '=4.3.3'
-ipycanvas = "^0.5.1"
-ipyevents = "^0.8.0"
-ipywidgets = "^7.5.1"
-```
+Ipyannotator provides a [simple API](https://palaimon.github.io/ipyannotator/nbs/21_api_doc.html) that provides the ability to explore, create and improve annotation datasets by using a pair of input/outputs. All pair of input/output are [listed on Ipyannotator's docs](https://palaimon.github.io/ipyannotator/nbs/22_input_output_doc.html). Check [Ipyannotator tutorials](https://palaimon.github.io/ipyannotator/docs/tutorials.html) for a quickly demonstration of the library.
 
-## Run ipyannotator as stand-alone web app using voila
+### Run ipyannotator tests
 
-Using `poetry`:
+The notebooks contains the development dependencies: [pytest](https://docs.pytest.org/en/7.1.x/) and [ipytest](https://github.com/chmp/ipytest). The tests can be executed using `nbdev_test_nbs` on the terminal. 
 
-install:
+### Run ipyannotator as stand-alone web app using voila
+
+Ipyannotator can be executed as a web app using the [voila](https://github.com/voila-dashboards/voila) library.
+
+#### Using poetry
+
+After the lib it's installed:
+
 ```shell
 cd {project_root}
 poetry install --no-dev
-poetry run pip install voila
 ```
-and run simple ipyannotator standalone example:
-```shell 
+
+Any jupyter notebook with ipyannotator can be executed as an standalone web application. An example of voila usage it's available in the current repository and can be executed as it follow:
+
+```shell
 poetry run voila nbs/09_voila_example.ipynb --enable_nbextensions=True
 ```
-  
-Same with `pip`:
+
+#### Using pip
+
+The installation and execution process can also be done using pip.
 
 ```shell
    cd {project_root}
    
-   pip install . 
+   pip install .
    pip install voila
    
[…]
