### 2022-07-11 cd1639e
Update paper.md to add the 'infinitely hackable annotation framework' section
