### 2022-07-12 00d96b7
Update JOSS paper (#28)
* Update paper.md to add the 'infinitely hackable annotation framework' section

* Update paper.md and paper.bib
