+++ paper/paper.bib
@@ -95,6 +95,24 @@
   year = {2022}
 }
 
+@software{ipycanvas2022,
+  author = {Martin Renou},
+  month = {07},
+  title = {Ipycanvas},
+  url = {https://github.com/martinRenou/ipycanvas},
+  version = {0.12.0},
+  year = {2022}
+}
+
+@software{ipyevents2022,
+  author = {Matt Craig},
+  month = {07},
+  title = {Ipycanvas},
+  url = {https://github.com/mwcraig/ipyevents},
+  version = {0.12.0},
+  year = {2022}
+}
+
 @software{voila2022,
   author = {Voila Dashboards},
   month = {03},
