### 2022-07-12 d352c3a
Deploying to gh-pages from @ palaimon/ipyannotator@97bf194c3ef01357eb34dabe1cd8bd246cfeb7ce 🚀
