+++ _sources/nbs/01d_tutorial_video_annotator.ipynb.txt
@@ -325,7 +325,7 @@
  ],
  "metadata": {
   "kernelspec": {
-   "display_name": "Python 3",
+   "display_name": "Python 3 (ipykernel)",
    "language": "python",
    "name": "python3"
   }
