+++ _sources/nbs/01c_tutorial_bbox.ipynb.txt
@@ -351,7 +351,7 @@
  ],
  "metadata": {
   "kernelspec": {
-   "display_name": "Python 3",
+   "display_name": "Python 3 (ipykernel)",
    "language": "python",
    "name": "python3"
   }
