### 2022-07-12 #4480 @itepifanio
@csadorf thanks for your feedback. I believe that our last modifications address your suggestions.

You can find a PR with the [JOSS paper changes](https://github.com/palaimon/ipyannotator/pull/28) and another with the [documentation enhancements](https://github.com/palaimon/ipyannotator/pull/27). The following changes were made:

- We add a new section on the paper to describe why Ipyannotator is an "infinitely hackable" library
- The list of dependencies was removed from the README. The information was necessary in the past, but now all dependencies can be found on pyproject.toml
- The follow-up instructions to the install step were added on the README, and in the docs
- A better introduction (statement of need) was added to the docs
- The API documentation was fully elaborated, adding more resources, describing decisions and linking it with the tutorials
- A new section describing how to run the library tests was added to the docs

Please let us know if you have any other concerns
