### 2022-07-14 #4480 @csadorf
> @csadorf thanks for your feedback. I believe that our last modifications address your suggestions.
>
> You can find a PR with the [JOSS paper changes](https://github.com/palaimon/ipyannotator/pull/28) and another with the [documentation enhancements](https://github.com/palaimon/ipyannotator/pull/27). The following changes were made:
>

Thank you for addressing my concerns.

> * We add a new section on the paper to describe why Ipyannotator is an "infinitely hackable" library

I think the edited description provides a much better justification for this claim, however the fact that ipyannotator is implemented in Python and executed via Jupyter notebooks provides barely any justification for this claim since the same could be said for literally any other open-source software. While I understand that this is to be understood as hyperbole, I think it is important to point out very concretely which elements specifically make this software significantly more customizable compared to similar open-source frameworks.

I think both the paper and the docs are providing sufficient explanation now, the only suggestion I would make is to remove any language that appears to be imply that (Python + OSS) is sufficient grounds for the claim. Specifically I have an issue with this sentence:

`Since Ipyannotator runs on top of Jupyter notebooks and due to the dynamic nature of the Python language, extra features and customization can be easily developed by the data science team, creating an "infinitely hackable annotation framework".
`

> * The list of dependencies was removed from the README. The information was necessary in the past, but now all dependencies can be found on pyproject.toml

✅
> * The follow-up instructions to the install step were added on the README, and in the docs

✅

> * A better introduction (statement of need) was added to the docs

✅

> * The API documentation was fully elaborated, adding more resources, describing decisions and linking it with the tutorials

I still find it a bit hard to navigate the docs, but it looks like the most important elements are now present.

> * A new section describing how to run the library tests was added to the docs
>

I am glad to see that the concern was addressed, however there is some significant room for improvement here.

It is usually good practice to define a test environment explicitly. The docs only state that tests require `pytest` and `ipytest` without any version information or instructions on how to create the environment. That's odd because it appears that the test environment is well-defined as part of the `pyproject.toml` file?

Further, the instructions on how to run the tests remain weirdly vague as well. Looking at the CI workflows, it seems that the tests are run via poetry? Is this recommended or required? It is not entirely clear to me why the instructions remain so vague when at the same time tests are run rigorously on each commit as it seems.

> Please let us know if you have any other concerns

I don't have any other concerns but those mentioned above.
