+++ paper/paper.md
@@ -33,13 +33,14 @@ Ipyannotator is a library developed to provide a solution to this problem. Ipyan
 
 Jupyter is one of the most popular tools for data science [@wang2019how] and is currently used by more than 7850000 public repositories on GitHub [@parente2022nbestimate]. Ipyannotator is a tool developed to be used on Jupyter Notebook, allowing researchers and developers to integrate the library into their ML projects quickly and easily. The solution, however, is not limited to research and developement teams. Since the Ipyannotator runs also on a web server, it can by used for annotation purposes by any user.
 
-[@lahtinen2021brima] developed an annotator as a browser plugin, [@dutta2019via] built an annotator as a program that runs on the browser, and [@bernal2019gtcreator] developed a desktop application for domain-specific medical image annotation. The previous annotation tools are restricted to fixed data types and a specific type of annotation. Since Ipyannotator runs on top of Jupyter notebooks and due to the dynamic nature of the Python language, extra features and customization can be easily developed by the data science team, creating an "infinitely hackable annotation framework".
+[@lahtinen2021brima] developed an annotator as a browser plugin, [@dutta2019via] built an annotator as a program that runs on the browser, and [@bernal2019gtcreator] developed a desktop application for domain-specific medical image annotation. The previous annotation tools are restricted to fixed data types and a specific type of annotation. Ipyannotator was designed to be executed on Jupyter notebook and allow users to change its iteractions on runtime, allowing extensions and customizations, creating an "infinitely hackable annotation framework" [^1].
 
-# Infinitely hackable annotation framework
+[^1]: The "infinitely hackable" claim emerged from Ipyannotator be significantly more customizable than similar open-source  libraries.
 
+# Infinitely hackable annotation framewor
 Ipyannotator uses Python libs, such as Ipywidgets [@ipywidgets2022], Ipycanvas [@ipycanvas2022] and Ipyevents [@ipyevents2022], that abstract the HTML and Javascript interactions, allowing developers to design UI interactions and elements using Python. Python's dynamic nature allows users to modify classes or modules in runtime, due to the libs mentioned, users can change the default behavior of Ipyannotator's UI, hacking the library. Browser interaction like mouse moving and HTML elements like dropdowns are some of the examples that can be changed at runtime when using Ipyannotator.
 
-Being integrated with Jupyter notebooks makes Ipyannotator usage easy to modify at different abstraction levels. The data science team can change the library behavior while writing their own scripts on a platform and using a programming language that they already know.
+Being integrated with Jupyter notebooks makes Ipyannotator usage easy to modify at different abstraction levels. The data science team can change the library behavior while writing their own scripts on a platform and programming language that they already know.
 
 In addition to the ability to change Ipyannotator's browser interaction using Python, the library also provides a flexible API that enables adding custom annotators. With a custom pair of input and output classes, the user can create and register a new annotator while reusing the library resources.
 
