### 2022-07-15 add6d35
Add better description to the tests (#31)
* Add better description to the tests

* Update README.md
