### 2022-07-16 a4da6d2
Improve infinitely hackable claim (#30)
* explain infinitely hackable phrase
