### 2022-07-18 #4480 @itepifanio
@csadorf thanks for the suggestions.

We tried to make the "infinitely hackable" claim clearer on [#30](https://github.com/palaimon/ipyannotator/pull/30).

[#31](https://github.com/palaimon/ipyannotator/pull/31) tries to improve the documentation of the tests.

>  Is this recommended or required?

We're running using poetry only because it's easier on the CI (it guarantees that the environment is active), it's not a requirement or recommendation.
