### 2022-07-19 palaimon/ipyannotator#33 @csadorf
Make instructions on how to run tests explicit
_This issue was created in response to [this comment](https://github.com/openjournals/joss-reviews/issues/4480#issuecomment-1187445385) as part of the [JOSS paper review](https://github.com/openjournals/joss-reviews/issues/4480)._

> [#31](https://github.com/palaimon/ipyannotator/pull/31) tries to improve the documentation of the tests.

Unfortunately the instructions remain ambiguous. Instructions on how to run tests should be as unambiguous and explicit as possible. I will explain what I mean by quoting from the improved instructions:

---

> When installing the repository using poetry, all dev dependencies are installed by default.

Instructions on how to install via poetry only come later and include the `--no-dev` option – meaning that test dependencies are in fact not installed unless one infers that this option must be dropped.

> When using pip for installation make sure to install the two dev dependencies `pytest` and `ipytest`, with the versions listed in `pyproject.toml`, manually:
>
> ```shell
> pip install pytest
> pip install ipytest
> ```

The instructions and the shown commands contradict each other since the latter will always install the latest version of the respective packages. Further,  a version for `pytest` **is not listed** in the `pyproject.toml` file.

---

It is _ok_ to be opinionated about how tests are supposed to be run for your package. If creating the correct environment is easier done using `poetry` then simply make that the recommended path, you can always provide instructions for alternative paths that can then be slightly less explicit. Here is my take at the instructions:

> To run tests:
> 1. [Install poetry](https://python-poetry.org/docs/#installation)
> 2. Create the test environment with `$ poetry install`
> 3. Run tests by executing `$ nbdev_test_nbs`
