### 2022-07-19 #4480 @csadorf
> @csadorf thanks for the suggestions.
>
> We tried to make the "infinitely hackable" claim clearer on [#30](https://github.com/palaimon/ipyannotator/pull/30).
>

I think it is sufficient now.

> [#31](https://github.com/palaimon/ipyannotator/pull/31) tries to improve the documentation of the tests.

Unfortunately the instructions remain ambiguous. I have provided a detailed explanation in [this issue](https://github.com/palaimon/ipyannotator/issues/33).
