### 2022-07-25 cdcfba2
Update test section (#34)
* Update README.md
