### 2022-07-25 #4480 @csadorf
> @csadorf [palaimon/ipyannotator#34](https://github.com/palaimon/ipyannotator/pull/34) has been merged now. Does this now complete your last open checkbox? Please let us know if you have further requests before we get your final review approval.

All good from my side.
