### 2022-08-09 #4480 @matthewfeickert
> Hi @matthewfeickert - coincidentally, I've been on vacation for 2 weeks and am just checking in on this, and I see it's the date that you said you will have some time smile

Yup. Hope you had a good holiday (and that reentry isn't too brutal)! :) JOSS reviews are the subject of my evening after work today. :+1:
