### 2022-08-11 palaimon/ipyannotator#37 @matthewfeickert
[JOSS Review] Add automated tests in CI
For the JOSS documentation check of https://github.com/openjournals/joss-reviews/issues/4480

> **Automated tests**: Are there automated tests or manual steps described so that the functionality of the software can be verified?

There are no [automated tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) (there are [GitHub Action workflows for building and deploying the docs](https://github.com/palaimon/ipyannotator/blob/cdcfba205d0fdf72c49127146f7a5586ee97970c/.github/workflows/docs_pages.yaml), but not for running the tests across all versions of Python supported). It seems that CI with some basic coverage of the API would greatly improve things.

For `v0.8.3` there are no instructions on how to run the tests in the `README` (but they were added in for a future version in PR #34). It is additionally unclear what the test coverage of running the tests in the notebooks is, which would be important to know.
