### 2022-08-11 palaimon/ipyannotator#38 @matthewfeickert
[JOSS Review] Suggested revisions on code quality
For suggested revisions for https://github.com/openjournals/joss-reviews/issues/4480:

In the code there are multiple instances of things like

```python
from nbdev import *
```

```console
$ git grep 'import \*'
dev_notes.md:from nbdev.export import *
nbs/00d_doc_utils.ipynb:    "from nbdev import *"
nbs/01_bbox_canvas.ipynb:    "from nbdev import *"
nbs/01_helpers.ipynb:    "from nbdev import *\n",
nbs/01a_datasets.ipynb:    "from nbdev import *"
nbs/01b_dataset_video.ipynb:    "from nbdev import *"
nbs/01b_tutorial_image_classification.ipynb:    "from nbdev import *"
nbs/01c_tutorial_bbox.ipynb:    "from nbdev import *"
nbs/01d_tutorial_video_annotator.ipynb:    "from nbdev import *"
nbs/02_navi_widget.ipynb:    "from nbdev import *\n",
nbs/02a_right_menu_widget.ipynb:    "from nbdev import *"
nbs/02b_grid_menu.ipynb:    "from nbdev import *"
nbs/03_storage.ipynb:    "from nbdev import *\n",
nbs/04_bbox_annotator.ipynb:    "from nbdev import *"
nbs/05_image_button.ipynb:    "from nbdev import *"
nbs/06_capture_annotator.ipynb:    "from nbdev import *"
nbs/07_im2im_annotator.ipynb:    "from nbdev import *"
nbs/13_datasets_legacy.ipynb:    "from nbdev import *\n",
nbs/15_coordinates_input.ipynb:    "from nbdev import *"
nbs/16_custom_buttons.ipynb:    "from nbdev import *"
nbs/18_bbox_trajectory.ipynb:    "from nbdev import *\n",
nbs/19_bbox_video_annotator.ipynb:    "from nbdev import *"
scripts/check_lint.sh:# F403 -> 'from module import *' used; unable to detect undefined names
scripts/check_lint.sh:# Cause : "from nbdev import *"
```

The use of `import *` should be avoided at all costs in Python code unless there is an extremely good reason to use it. For anything close to a library I do not think there are any.
