### 2022-08-11 palaimon/ipyannotator#39 @matthewfeickert
[JOSS Review] Suggested revisions on paper quality
For suggested revisions for https://github.com/openjournals/joss-reviews/issues/4480:

* The paper is too long and JOSS papers are not meant to describe the API. Much of the section "A simple but flexible API to define annotation tasks" should be moved to the documentation, where it would be quite useful and welcome, though a simplified summary of the explore, create, improve workflow can remain in the paper. "Key Design Decisions" should also be cut and moved to a section of the documentation as well. It would probably be helpful to reread the [Submitting a paper to JOSS](https://joss.readthedocs.io/en/latest/submitting.html#submitting-a-paper-to-joss) instructions on what the paper should cover.

* In the Acknowledgements section you mention

> The authors acknowledge the financial support by the Federal Ministry for Digital and Transport of Germany under under the program mFUND for the project OS-VAT (project number 19F2160A).

Congratulations on the funding! If possible can you also include a reference or hyperlink to the project funding page? I'm not sure if there is another page with your actual propsal of if https://www.bmvi.de/SharedDocs/DE/Artikel/DG/mfund-projekte/os-vat.html is the actual government page for it.
