### 2022-08-11 #4480 @matthewfeickert
Thanks for your submission and software @itepifanio. I normally paste my full review based on the checklist here, but to reduce the length of the thread I am going to just link to my review checklist comment: https://github.com/openjournals/joss-reviews/issues/4480#issuecomment-1156588361. As always, the review is following the [JOSS paper review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html). Also, please feel free to ask for clarification on anything as I'd like to make sure that this information is helpful and actionable to you. :+1:

For consistency and reproducibility of all my comments, I will be using the public Docker image `python:3.10` to conduct my review which has the following software specs

```console
$ docker run --rm -ti -v $PWD:$PWD -w $PWD python:3.10 /bin/bash
# python --version --version
Python 3.10.6 (main, Aug  3 2022, 10:13:24) [GCC 10.2.1 20210110]
```

In addition to the Issues that I have already mentioned there I'll also include the following suggestions/summary

## Suggested revisions

### Code
* Automated tests should be added that cover all the versions of Python stated to be supported. Installation dependencies should be better clarified and a rough scope of coverage of the API should be states for the tests. Issue https://github.com/palaimon/ipyannotator/issues/36 addresses this.
* https://github.com/palaimon/ipyannotator/issues/38

#### Paper

* https://github.com/palaimon/ipyannotator/issues/39
