### 2022-08-11 #4480 @matthewfeickert
I also forgot to mention that I reviewed [`ipyannotator` `v0.8.3`](https://github.com/palaimon/ipyannotator/tree/v0.8.3) and from the other comments it seems that @csadorf did the same (or with `v0.8.2`). However, the review is listed for `v0.8.1`. Can we have `editorialbot` update the review version listed in https://github.com/openjournals/joss-reviews/issues/4480#issue-1272281133?
