### 2022-08-12 53dc4e2
Update setup.py to make requirements more clear
