+++ setup.py
@@ -3,9 +3,10 @@ from configparser import ConfigParser
 import setuptools
 assert parse_version(setuptools.__version__)>=parse_version('36.2')
 
-# note: all settings are in settings.ini; edit there, not here
+# note: most of the settings are in settings.ini; edit there, not here
+# the remain settings are in pyproject.toml (basically the requirements)
 config = ConfigParser(delimiters=['='])
-config.read('settings.ini')
+config.read(['settings.ini', 'pyproject.toml'])
 cfg = config['DEFAULT']
 
 cfg_keys = 'version description keywords author author_email'.split()
@@ -18,9 +19,9 @@ licenses = {
 }
 statuses = [ '1 - Planning', '2 - Pre-Alpha', '3 - Alpha',
     '4 - Beta', '5 - Production/Stable', '6 - Mature', '7 - Inactive' ]
-py_versions = '2.0 2.1 2.2 2.3 2.4 2.5 2.6 2.7 3.0 3.1 3.2 3.3 3.4 3.5 3.6 3.7 3.8'.split()
+py_versions = '3.8 3.9 3.10'.split()
 
-requirements = cfg.get('requirements','').split()
+requirements = (config['tool.poetry.dependencies'] or '').split()
 lic = licenses[cfg['license']]
 min_python = cfg['min_python']
 
@@ -43,4 +44,3 @@ setuptools.setup(
     zip_safe = False,
     entry_points = { 'console_scripts': cfg.get('console_scripts','').split() },
     **setup_cfg)
-
