+++ settings.ini
@@ -10,7 +10,7 @@ author_email = oss@mail.palaimon.io
 copyright = Palaimon GmbH
 branch = main
 version = 0.8.3
-min_python = 3.7
+min_python = 3.8
 audience = Developers
 language = English
 # Set to True if you want to create a more fancy sidebar.json than the default
@@ -20,8 +20,8 @@ license = apache2
 # From 1-7: Planning Pre-Alpha Alpha Beta Production Mature Inactive
 status = 2
 
-requirements = traitlets==4.3.3 jupyterlab>=2.2.8 ipywidgets>=7.5.1 ipyevents>=0.8.0 ipycanvas>=0.5.1
-dev_requirements = nbdev RISE nbdime
+#requirements = traitlets==4.3.3 jupyterlab>=2.2.8 ipywidgets>=7.5.1 ipyevents>=0.8.0 ipycanvas>=0.5.1
+#dev_requirements = nbdev RISE nbdime
 
 # Optional. Same format as setuptools console_scripts
 # console_scripts = 
