### 2022-08-13 palaimon/ipyannotator#38 @itepifanio
Hi @matthewfeickert.

We are not a big fan of the `import *` heavily used and advocated by nbdev [0, 1] (even though we like many of the unconventional ideas used in nbdev).
Given that nbdev was designed explicitly to work with `import *` and using explicit imports with this library is not very convenient we decided to use a compromise.

- the dev notebooks use the `import *`, but only for nbdev modules
- our library doesn't use any `import *`, so our users don't have to face any of the potential downsides of `import *`.

As you can see from your grep these imports are located on the notebooks only and are not exported to the final library (located in the ipyannotator folder).
Strictly speaking, our library doesn't use any `import *`, can you live with the compromise we explained above?

If not how would you suggest to proceed?

[0] Their [boilerplate code](https://nbdev1.fast.ai/example.html#Edit-index.ipynb) already comes with the `import *` for example.
[1] hn discussion with nbdev core dev https://news.ycombinator.com/item?id=24386147
