### 2022-08-13 palaimon/ipyannotator#37 @itepifanio
@matthewfeickert

Thanks for the comment. I think this requires a nbdev specific explanation as well. :)

### Tests in ci
This is something we have also discussed with our other reviewer. We [had first included an explanation into the docs](https://github.com/palaimon/ipyannotator/pull/34) but have removed it again later for clarity.

In short, nbdev doesn't separate the tests from the notebooks. When the CI builds the notebooks code it's also running the tests. From nbdev docs:

"Everything that is not an exported cell is considered a test, so you should make sure your notebooks can all run smoothly (and fast) if you want to use this functionality as the CLI."

The tests were written in notebooks right next to the doc and code. We're using ipytests to add tests in the notebooks and the assert keyword for simple tests.

### Coverage
There [are a few issues](https://forums.fast.ai/t/nbdev-code-coverage-n-tests/73993) when trying to run coverage with nbdev, but we're going to investigate how to do it and come up with at least a rough number. In addition to the technical difficulty to get the number, ipyannotator has a fair amount of ui-related code which is difficult to automatically test (eyeballing notebooks with rendered widgets is unreasonable effective) which makes the line coverage score even harder to interpret than it usually is.
