### 2022-08-15 085c150
Update dependencies and python version on pyproject.toml
