+++ pyproject.toml
@@ -8,14 +8,14 @@ readme = "README.md"
 repository = "https://github.com/palaimon/ipyannotator"
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = ">=3.8,<3.11"
 pandas = "^1.1.2"
 traitlets = "^5.1.1"
 ipycanvas = "^0.10.2"
 ipyevents = "^0.8.0"
 ipykernel = "^5.3.4"
 ipywidgets = "^7.5.1"
-scikit-image = "^0.18.3"
+scikit-image = "^0.19.3"
 matplotlib = "^3.4.3"
 pydantic = "^1.8.2"
 PyPubSub = "^4.0.3"
