+++ settings.ini
@@ -20,9 +20,6 @@ license = apache2
 # From 1-7: Planning Pre-Alpha Alpha Beta Production Mature Inactive
 status = 2
 
-#requirements = traitlets==4.3.3 jupyterlab>=2.2.8 ipywidgets>=7.5.1 ipyevents>=0.8.0 ipycanvas>=0.5.1
-#dev_requirements = nbdev RISE nbdime
-
 # Optional. Same format as setuptools console_scripts
 # console_scripts = 
 
