### 2022-08-15 d869e2e
Add CI tests for multiple python versions
