+++ .github/workflows/release.yaml
@@ -10,11 +10,14 @@ jobs:
   all_tests:
     name: "Run NB Tests"
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.8", "3.9", "3.10"]
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v3
+      - uses: actions/setup-python@v4
         with:
-          python-version: '3.8'
+          python-version: ${{ matrix.python-version }}
       - uses: abatilo/actions-poetry@v2.1.4
         with:
           poetry-version: 1.1.0
@@ -33,7 +36,6 @@ jobs:
       - name: Run Notebooks tests
         run: poetry run nbdev_test_nbs
 
-
   release_whl:
     needs: all_tests
     if: github.event_name == 'release'
@@ -41,7 +43,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v3
+      - uses: actions/setup-python@v4
       - uses: abatilo/actions-poetry@v2.1.4
         with:
             poetry-version: 1.1.0
