### 2022-08-16 palaimon/ipyannotator#37 @ibayer
test coverage as reported by  #41

```
rootdir: /home/italo/workspace/github_ipyannotator
plugins: anyio-3.5.0, cov-3.0.0
collected 1 item

test_nbs.py .                                                            [100%]

---------- coverage: platform linux, python 3.8.10-final-0 -----------
Name                                         Stmts   Miss  Cover
----------------------------------------------------------------
ipyannotator/__init__.py                         1      0   100%
ipyannotator/_nbdev.py                           6      0   100%
ipyannotator/annotator.py                      111      5    95%
ipyannotator/base.py                            83     12    86%
ipyannotator/bbox_annotator.py                 187      3    98%
ipyannotator/bbox_canvas.py                    379     81    79%
ipyannotator/bbox_video_annotator.py           272      8    97%
ipyannotator/capture_annotator.py              183      6    97%
ipyannotator/custom_input/__init__.py            0      0   100%
ipyannotator/custom_input/buttons.py            98      0   100%
ipyannotator/custom_input/coordinates.py        52      8    85%
ipyannotator/custom_widgets/__init__.py          0      0   100%
ipyannotator/custom_widgets/grid_menu.py        86      7    92%
ipyannotator/datasets/__init__.py                0      0   100%
ipyannotator/datasets/download.py               95     10    89%
ipyannotator/datasets/factory.py                48     26    46%
ipyannotator/datasets/factory_legacy.py         65     18    72%
ipyannotator/datasets/generators.py            301      6    98%
ipyannotator/datasets/generators_legacy.py     141     12    91%
ipyannotator/debug_utils.py                     61      5    92%
ipyannotator/docs/__init__.py                    0      0   100%
ipyannotator/docs/api.py                         9      0   100%
ipyannotator/docs/input_output.py               13      0   100%
ipyannotator/docs/utils.py                     142     51    64%
ipyannotator/explore_annotator.py               84      6    93%
ipyannotator/helpers.py                        152      5    97%
ipyannotator/im2im_annotator.py                217     35    84%
ipyannotator/ipytyping/__init__.py               0      0   100%
ipyannotator/ipytyping/annotations.py           72      5    93%
ipyannotator/mltypes.py                         58      1    98%
ipyannotator/navi_widget.py                     71      5    93%
ipyannotator/right_menu_widget.py              147      1    99%
ipyannotator/services/__init__.py                0      0   100%
ipyannotator/services/bbox_trajectory.py        31      1    97%
ipyannotator/storage.py                        273     21    92%
----------------------------------------------------------------
TOTAL                                         3438    338    90%

======================== 1 passed in 265.57s (0:04:25) =========================
```
