+++ paper/paper.md
@@ -51,12 +51,6 @@ These three steps in conjuction with domain-specific annotation types, define th
 
 The following code examples illustrate the main actions around which the Ipyannotator API is built. Please keep in mind that Ipyannotator aims to be flexible enought so that these generic aspects can be extented to much complexer and domain-specific tasks and interfaces. As an exemplary application, a standard image classification task, is used in the following.
 
-To set up Ipyannotator in Jupyter Notebook the library has to be imported with three main components: the pair of input/output data, the settings structure, and the generic annotator module. The following example uses the 
-```python 
-InputImage/OutputImageLabel
-```
-associated with the image classification task. It imports the settings pointing to a folder called 'data' and configures the annotator module with the entries. Once the annotator module is configured it can be used to call the explore, create and improve steps.
-
 ```python
 from pathlib import Path
 from ipyannotator.base import Settings
@@ -74,21 +68,10 @@ annotator = Annotator(input, output, settings)
 The explore step provides the ability to visualize and navigate through images and datasets. \autoref{fig:image-classification-explore} displays an example of an image classification task using one of Ipyannotator's inbuilt artificial demonstration datasets.
 The dataset consists of images showing simple shapes in different colors. The annotation task is to assign to each image the color which is closest to the provided labels.
 
-The explore step is started by the command:
-
-```python
-annotator.explore()
-```
 ![Explore step for image classification task\label{fig:image-classification-explore}](image-classification-explore.png){ width=60% }
 
 ## Create
 In the create step new annotation datasets can be created by the user. \autoref{fig:image-classification-create} presents an example of an image classfication task, allowing the user to manually select multiple options and save the results in the dataset.
-
-The create step is started by the following instruction:
-
-```python
-annotator.create()
-```
 
 ![Create step for image classification task \label{fig:image-classification-create}](image-classification-create.png)
 
@@ -96,10 +79,6 @@ annotator.create()
 The improve step enables the user to refine datasets and, thereby, improve the annotation quality. \autoref{fig:image-classification-improve} displays the improve usage, allowing users to iterate over every class and identify incorrect results.
 
 Inspecting large batches of pre-annotated data allows to very quickly improve the quality of datasets that were generated by an insufficient ML model. It also allows a domain expert to verify and improve the annotation work of less specialized annotators.
-
-```python
-annotator.improve()
-```
 
 ![Improve step for image classification task \label{fig:image-classification-improve}](image-classification-improve.png){ width=60% }
 
@@ -119,38 +98,6 @@ Ipyannotator's architecture consists of three main systems components that compr
 ![Technical user architecture  \label{fig:technical-user-architecture}](technical-user-architecture.png){ width=52% }
 
 For the technical user multiple tutorials are provided, demonstrating Ipyannotator's utilization. The tutorials make it easier for new users get started and adapt the notebooks to their tasks. They also demonstrate annotation workflow and different features.
-
-## Layers
-The layering design leads to loosely coupled systems, isolating responsibilities and allows to easier implement changes [@gamma1995design]. Ipyannotator is separated into multiple layers to provide an architecture that allows the annotator to easier add extensions. The library uses four layers: view, controller, storage, and state. Each layer can be rewritten, extended and customized.
-
-- The *View* layer is responsible for rendering the visualizations. Ipyannotator uses ipycanvas and ipywidgets to structure and mount the visualization layer. Additionally, internal components such as the navigation menue were developed which helps the users to navigate through the images that need to be annotated.
-
-- The *Storage* layer is the layer that receives the data and stores it. Ipyannotator uses different types of storage formats like .txt and .SQLite.
-
-- The *Controller* layer acts as a mediator between state, storage and view. This layer tells when the information from the state will be stored.
-
[…]
