### 2022-08-19 22928a8
Fix typo. Remove unused png files
