### 2022-08-19 #4480 @itepifanio
@matthewfeickert the [paper-related issue](https://github.com/palaimon/ipyannotator/issues/39) was addressed on [PR 42](https://github.com/palaimon/ipyannotator/pull/42) and [PR 43](https://github.com/palaimon/ipyannotator/pull/43).
