### 2022-08-22 palaimon/ipyannotator#37 @matthewfeickert
> In short, nbdev doesn't separate the tests from the notebooks. When the CI builds the notebooks code it's also running the tests. From nbdev docs:
>
> "Everything that is not an exported cell is considered a test, so you should make sure your notebooks can all run smoothly (and fast) if you want to use this functionality as the CLI."
>
> The tests were written in notebooks right next to the doc and code. We're using ipytests to add tests in the notebooks and the assert keyword for simple tests.

Yes, I am familiar with nbdev as I already mentioned but this is an Issue about lacking CI, which at the moment is not addressed in PR #41 at all.

90% coverage is rather low, but I will consider this as acceptable for publication. I would strongly suggest that you bring your coverage up.
