### 2022-08-22 palaimon/ipyannotator#38 @matthewfeickert
> As you can see from your grep these imports are located on the notebooks only and are not exported to the final library (located in the ipyannotator folder).
> Strictly speaking, our library doesn't use any `import *`, can you live with the compromise we explained above?

That's valid! I appreciate the compromise that you've already made in your approaches. :+1:

> If not how would you suggest to proceed?

Sorry, I didn't make this clear enough, but you don't need to address this at all to proceed with the review. This is just a suggestion that I am making as someone who has poked around you code during the review, but it is not at all something that is keeping you from moving forward with publication. :+1:
