### 2022-08-22 #4480 @matthewfeickert
@danielskatz As I think `editorialbot` needs to be able to build from a branch on the target repo itself (as opposed to being able to build from a branch that is part of a PR to the repo) @itepifanio will need to merge https://github.com/palaimon/ipyannotator/pull/42 before we can see the paper render with those changes, correct?
