### 2022-08-22 palaimon/ipyannotator#37 @matthewfeickert
> There are no [automated tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) (there are [GitHub Action workflows for building and deploying the docs](https://github.com/palaimon/ipyannotator/blob/cdcfba205d0fdf72c49127146f7a5586ee97970c/.github/workflows/docs_pages.yaml), but not for running the tests across all versions of Python supported). It seems that CI with some basic coverage of the API would greatly improve things.

This statement of mine was false as of PR #2 given

https://github.com/palaimon/ipyannotator/blob/cdcfba205d0fdf72c49127146f7a5586ee97970c/.github/workflows/release.yaml#L33-L34

PR #40 improves on this by testing across all supported Python versions, and will resolve this Issue. :+1:

> test coverage as reported by #41

90% coverage is rather low, but I will consider this as acceptable for publication. :+1:  I would strongly suggest that you bring your coverage up.
