### 2022-08-22 palaimon/ipyannotator#37 @matthewfeickert
(There was a previous comment from me that I deleted that made an incorrect statement about there needing to be additional work. Please ignore that comment in any email notifications you received.)
