### 2022-08-22 palaimon/ipyannotator#37 @ibayer
I'm closing the issue since the confusion about missing CI has been resolved and
the suggestion to improve coverage is tracked elsewhere.
