### 2022-08-22 006edb0
Paper quality enhancements (#42)
* Update paper.md

* Fix typo. Remove unused png files
