### 2022-08-22 palaimon/ipyannotator#39 @ibayer
Thanks for the comments!

> if https://www.bmvi.de/SharedDocs/DE/Artikel/DG/mfund-projekte/os-vat.html is the actual government page for it.

This is indeed the official government page.

I'm closing this issue since the other comments have been addressed in the linked MRs.
