### 2022-08-22 ddcd475
Deploying to gh-pages from @ palaimon/ipyannotator@d44caf83b41eee1cec801f7b3adb6de373bb4757 🚀
