### 2022-08-22 #4480 @ibayer
@matthewfeickert

> @itepifanio will need to merge https://github.com/palaimon/ipyannotator/pull/42 before we can see the paper render with those changes, correct?

I have just merged the PRs you reviewed incl. https://github.com/palaimon/ipyannotator/pull/42.
