+++ .github/workflows/release.yaml
@@ -12,7 +12,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v3
+      - uses: actions/setup-python@v4.1.0
         with:
           python-version: '3.8'
       - uses: abatilo/actions-poetry@v2.1.4
@@ -41,7 +41,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v3
+      - uses: actions/setup-python@v4.1.0
       - uses: abatilo/actions-poetry@v2.1.4
         with:
             poetry-version: 1.1.0
