+++ .github/workflows/docs_pages.yaml
@@ -11,7 +11,7 @@ jobs:
     steps:
     - name: Checkout
       uses: actions/checkout@v3
-    - uses: actions/setup-python@v3
+    - uses: actions/setup-python@v4.1.0
       with:
         python-version: '3.8'
     - uses: abatilo/actions-poetry@v2.1.4
