### 2022-08-22 9693e38
Bump abatilo/actions-poetry from 2.1.4 to 2.1.5
Bumps [abatilo/actions-poetry](https://github.com/abatilo/actions-poetry) from 2.1.4 to 2.1.5.
- [Release notes](https://github.com/abatilo/actions-poetry/releases)
- [Changelog](https://github.com/abatilo/actions-poetry/blob/master/.releaserc)
- [Commits](https://github.com/abatilo/actions-poetry/compare/v2.1.4...v2.1.5)

---
updated-dependencies:
- dependency-name: abatilo/actions-poetry
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>
