+++ .github/workflows/release.yaml
@@ -15,7 +15,7 @@ jobs:
       - uses: actions/setup-python@v4.1.0
         with:
           python-version: '3.8'
-      - uses: abatilo/actions-poetry@v2.1.4
+      - uses: abatilo/actions-poetry@v2.1.5
         with:
           poetry-version: 1.1.0
       - name: Install ipyannotator
@@ -42,7 +42,7 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4.1.0
-      - uses: abatilo/actions-poetry@v2.1.4
+      - uses: abatilo/actions-poetry@v2.1.5
         with:
             poetry-version: 1.1.0
       - name: Upload
