### 2022-08-22 f7c2fc1
Update pyproject.toml
Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
