+++ pyproject.toml
@@ -8,7 +8,7 @@ readme = "README.md"
 repository = "https://github.com/palaimon/ipyannotator"
 
 [tool.poetry.dependencies]
-python = ">=3.8,<3.11"
+python = ">=3.8"
 pandas = "^1.1.2"
 traitlets = "^5.1.1"
 ipycanvas = "^0.10.2"
