+++ .github/workflows/release.yaml
@@ -43,8 +43,8 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
-      - uses: abatilo/actions-poetry@v2.1.4
+      - uses: actions/setup-python@v4.1.0
+      - uses: abatilo/actions-poetry@v2.1.5
         with:
             poetry-version: 1.1.0
       - name: Upload
