+++ pyproject.toml
@@ -9,13 +9,13 @@ repository = "https://github.com/palaimon/ipyannotator"
 
 [tool.poetry.dependencies]
 python = ">=3.8"
-pandas = ">=1.1.2"
+pandas = ">=1.2.0"
 traitlets = ">=5.1.1"
 ipycanvas = ">=0.10.2"
 ipyevents = ">=0.8.0"
 ipykernel = ">=5.3.4"
 ipywidgets = ">=7.6.0"
-scikit-image = ">=0.18.3"
+scikit-image = ">=0.19.3"
 matplotlib = ">=3.4.3"
 pydantic = "1.8.*"
 PyPubSub = "4.0.*"
