+++ .github/workflows/release.yaml
@@ -15,10 +15,10 @@ jobs:
         python-version: ["3.8", "3.9", "3.10"]
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4
+      - uses: actions/setup-python@v4.1.0
         with:
           python-version: ${{ matrix.python-version }}
-      - uses: abatilo/actions-poetry@v2.1.4
+      - uses: abatilo/actions-poetry@v2.1.5
         with:
           poetry-version: 1.1.0
       - name: Install ipyannotator
