### 2022-08-22 6e01e8c
Python version support (#40)
* Update setup.py to make requirements more clear

* Update dependencies and python version on pyproject.toml

* Remove comment

* Add CI tests for multiple python versions

* Update pyproject.toml

Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>

* Remove dependency restrictions by setting minimum version

* Fix conflict

* Fix ci

* Rollback downgrade at release.yaml

Co-authored-by: Matthew Feickert <matthew.feickert@cern.ch>
