+++ .github/workflows/release.yaml
@@ -10,11 +10,14 @@ jobs:
   all_tests:
     name: "Run NB Tests"
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.8", "3.9", "3.10"]
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4.1.0
         with:
-          python-version: '3.8'
+          python-version: ${{ matrix.python-version }}
       - uses: abatilo/actions-poetry@v2.1.5
         with:
           poetry-version: 1.1.0
@@ -32,7 +35,6 @@ jobs:
 
       - name: Run Notebooks tests
         run: poetry run nbdev_test_nbs
-
 
   release_whl:
     needs: all_tests
