### 2022-08-23 #4480 @matthewfeickert
I've updated my review in the checklist above https://github.com/openjournals/joss-reviews/issues/4480#issuecomment-1156588361. Given the comments I made there, I think that once https://github.com/palaimon/ipyannotator/pull/43#discussion_r952149168 is resolved that I am able to recommend moving forward with publication. :+1:
