### 2022-08-23 #4480 @danielskatz
@itepifanio - even though https://github.com/palaimon/ipyannotator/pull/43 is merged, @matthewfeickert's [comment in it](https://github.com/palaimon/ipyannotator/pull/43#discussion_r952149168) (after it was merged) seems like the last issue to address.
