+++ index.rst
@@ -15,3 +15,4 @@
    docs/internal-components
    ../nbs/22_input_output_doc
    ../nbs/21_api_doc
+   docs/design
