### 2022-08-23 #4480 @ibayer
> https://github.com/palaimon/ipyannotator/pull/43#discussion_r952149168 (after it was merged) seems like the last issue to address.

The requested fix has been merged now with https://github.com/palaimon/ipyannotator/pull/45 as well.
