### 2022-08-23 #4480 @ibayer
We are currently preparing a new release that contains the improvements from the review process as mentioned in https://github.com/openjournals/joss-reviews/issues/4480#issuecomment-1212557726 before. We intent to cut the release as soon as the review as been officially completed.
