### 2022-08-23 #4480 @danielskatz
A new release addressing changes from the process is a requirement for publication. You might want to wait slightly, as I'll be proofreading the paper as well once @matthewfeickert is satisfied, and any changes could also potentially be captured in the publication release
