### 2022-08-23 b8fb4c3
minor paper updates
