+++ paper/paper.md
@@ -20,36 +20,36 @@ bibliography: paper.bib
 ---
 
 # Summary
-Annotation is a task that associates semantic tags to a digital representation (image, video, text, and others). This process is important for supervised machine learning (ML) approaches, since the model learns and successively improves from data examples in form of input-output-pairs.
+Annotation is a task that associates semantic tags to a digital representation (image, video, text, and others). This process is important for supervised machine learning (ML) approaches, since the model learns, and successively improves, from data examples in form of input-output pairs.
 
-The variety of digital representations makes it difficult to develop a tool that is flexible and meets all requirements for machine learning applications in different projects. Ipyannotator is an open-source tool that allows manual annotation on multiple data formats, enabling researchers/users to explore, create and improve their datasets without leaving their own development environment, and empowering them to extend and customize the annotation process.
+The variety of digital representations makes it difficult to develop a tool that is flexible and meets all requirements for machine learning applications in different projects. Ipyannotator is an open-source tool that allows manual annotation in multiple data formats, enabling researchers/users to explore, create, and improve their datasets without leaving their own development environment, and empowering them to extend and customize the annotation process.
 
 # Statement of need
-Many breakthroughs in machine learning (ML) applications such as image classification, text understanding, and recommender systems belong to the class of supervised machine learning. These ML methods often require an extensive basis of annotated data from which the model learns. The amount and quality of the annotated data is essential to generate a model which yields accurate prediction [@wong2015smartannotator].
+Many breakthroughs in machine learning (ML) applications such as image classification, text understanding, and recommender systems belong to the class of supervised machine learning. These ML methods often require an extensive basis of annotated data from which the model learns. The amount and quality of the annotated data is essential to generate a model that yields accurate prediction [@wong2015smartannotator].
 
-The variety of annotation taks, data formats, and the respective visualization of data is enormous. Dealing with multiple domains of supervised ML, the large variety of applications is a challenging task considering that the tooling is potentially not flexible enough which imposes limitations to the user.
+The variety of annotation taks, data formats, and the respective visualization of data is enormous. When dealing with multiple domains of supervised ML, the large variety of applications is a challenging task, especially considering that the tooling is potentially not flexible enough, which imposes limitations to the user.
 
-Ipyannotator is a library developed to provide a solution to this problem. Ipyannotator can be used to visualize, create, and improve datasets, providing a flexible solution that can be extended and customized by the user within the development environment Jupyter.
+Ipyannotator is a library developed to provide a solution to this problem. Ipyannotator can be used to visualize, create, and improve datasets, providing a flexible solution that can be extended and customized by the user within the Jupyter development environment.
 
-Jupyter is one of the most popular tools for data science [@wang2019how] and is currently used by more than 7850000 public repositories on GitHub [@parente2022nbestimate]. Ipyannotator is a tool developed to be used on Jupyter Notebook, allowing researchers and developers to integrate the library into their ML projects quickly and easily. The solution, however, is not limited to research and developement teams. Since the Ipyannotator runs also on a web server, it can by used for annotation purposes by any user.
[…]
