+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @inproceedings{wong2015smartannotator,
-  title={Smartannotator an interactive tool for annotating indoor rgbd images},
+  title={SmartAnnotator an interactive tool for annotating indoor {RGBD} images},
   author={Wong, Yu-Shiang and Chu, Hung-Kuo and Mitra, Niloy J},
   booktitle={Computer Graphics Forum},
   volume={34},
@@ -7,7 +7,7 @@
   pages={447--457},
   year={2015},
   organization={Wiley Online Library},
-  doi = {10.1145/3359141},
+  doi = {10.1111/cgf.12574},
 }
 
 @article{wang2019how,
