### 2022-08-23 #4480 @danielskatz
@itepifanio - please see https://github.com/palaimon/ipyannotator/pull/46 and https://github.com/palaimon/ipyannotator/pull/47 for some suggested changes in the paper and bib
