### 2022-08-23 7ed8956
Bump version to 0.8.4 in pyproject.toml
