### 2022-08-24 c0acf08
Update changelog for JOSS release (#48)
* Update changelog

* Bump version to 0.8.4 in pyproject.toml
