### 2022-08-24 9ccb16f
Bump abatilo/actions-poetry from 2.1.5 to 2.1.6 (#49)
Bumps [abatilo/actions-poetry](https://github.com/abatilo/actions-poetry) from 2.1.5 to 2.1.6.
- [Release notes](https://github.com/abatilo/actions-poetry/releases)
- [Changelog](https://github.com/abatilo/actions-poetry/blob/master/.releaserc)
- [Commits](https://github.com/abatilo/actions-poetry/compare/v2.1.5...v2.1.6)

---
updated-dependencies:
- dependency-name: abatilo/actions-poetry
  dependency-type: direct:production
  update-type: version-update:semver-patch
...

Signed-off-by: dependabot[bot] <support@github.com>

Signed-off-by: 
[…]
