+++ .github/workflows/release.yaml
@@ -18,7 +18,7 @@ jobs:
       - uses: actions/setup-python@v4.1.0
         with:
           python-version: ${{ matrix.python-version }}
-      - uses: abatilo/actions-poetry@v2.1.5
+      - uses: abatilo/actions-poetry@v2.1.6
         with:
           poetry-version: 1.1.0
       - name: Install ipyannotator
@@ -44,7 +44,7 @@ jobs:
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4.1.0
-      - uses: abatilo/actions-poetry@v2.1.5
+      - uses: abatilo/actions-poetry@v2.1.6
         with:
             poetry-version: 1.1.0
       - name: Upload
