+++ .github/workflows/docs_pages.yaml
@@ -14,7 +14,7 @@ jobs:
     - uses: actions/setup-python@v4.1.0
       with:
         python-version: '3.8'
-    - uses: abatilo/actions-poetry@v2.1.5
+    - uses: abatilo/actions-poetry@v2.1.6
       with:
         poetry-version: 1.1.0
     - name: install
