+++ .github/workflows/release.yaml
@@ -15,7 +15,7 @@ jobs:
         python-version: ["3.8", "3.9", "3.10"]
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4.1.0
+      - uses: actions/setup-python@v4.2.0
         with:
           python-version: ${{ matrix.python-version }}
       - uses: abatilo/actions-poetry@v2.1.6
@@ -43,7 +43,7 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
-      - uses: actions/setup-python@v4.1.0
+      - uses: actions/setup-python@v4.2.0
       - uses: abatilo/actions-poetry@v2.1.6
         with:
             poetry-version: 1.1.0
