+++ settings.ini
@@ -9,7 +9,7 @@ author = Palaimon
 author_email = oss@mail.palaimon.io
 copyright = Palaimon GmbH
 branch = main
-version = 0.8.4
+version = 0.8.5
 min_python = 3.8
 audience = Developers
 language = English
