+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "ipyannotator"
-version = "0.8.4"
+version = "0.8.5"
 description = "The infinitely hackable annotation framework"
 authors = ["palaimon.io <oss@mail.palaimon.io>"]
 license = "Apache License 2.0"
