### 2022-08-24 20195b9
Deploying to gh-pages from @ palaimon/ipyannotator@41858f4fbf545cf51a4a1bd5a3f2e5d8cae2da84 🚀
