### 2022-08-24 f41ca38
Joss review (#50)
* build new version

* add contributions

* bump version to avoid collision with internal release

Co-authored-by: Immanuel Bayer <git@ibayer.eu>
