### 2022-08-24 b2e384b
Deploying to gh-pages from @ palaimon/ipyannotator@36bd6eb75e7023e5c354fff0a9e8821d54b4d5bf 🚀
